requirements.txt
# requirements.txt - Restaurant Manager Pro (backend DRF)
# Phase 0 + Phase 1
# Commande : pip install -r requirements.txt

# Django
Django==5.1

# Django REST Framework - Phase 1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-spectacular==0.29.0

# Base de donnees PostgreSQL
psycopg[binary]==3.2.3

# Variables d'environnement
python-dotenv==1.0.1

# Images - Upload photos des plats
Pillow==10.4.0

# PDF - Generation recus et rapports (Phase 5 et 8)
reportlab==4.2.2

# QR Code - Connexion automatique tables (Phase 6)
qrcode[pil]==8.0

# Fichiers statiques - Production
whitenoise==6.7.0

# Serveur WSGI - Production
gunicorn==23.0.0

# Requests - Envoi d'emails via Resend
requests==2.32.3

# Resend - Envoi d'emails
resend==2.27.0

# --- A DECOMMENTER EN PHASE 8 ---
# celery==5.4.0
# django-celery-beat==2.7.0
# redis==5.2.0
# openpyxl==3.1.5